Fixed Content Type issue and EAs without a Script
I needed to add Accept: application/xml to the headers as the Jamf Pro server now serves up json by default. I had originally put in content-type: application/xml but I suppose that would only apply to uploads... I removed line 70 that contained a print statement that would cause the script to error if you have EAs without scripts. I was checking for the condition but if it matched the print would cause a failure. 

I added some light error handling on the first request to the Jamf Pro server. 

I also added some logic to prompt for a password if you omit the --password argument.