Multiple fixes to sync.py
This should fix #19 where update_all did not in fact, update all. The logic was flawed in that if there were no changes at all, the function would return prematurely. I rearranged the logic so it will upload all appropriately. 

Additionally, there was an issue with the headers not being explicitly set to accept xml. 

I added some additional logging

I added the option to omit the password and be prompted if no password is supplied on the command line.